requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation_models_pytorch>=0.3.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=9.5.0
albumentations>=1.3.0
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
